Made VGG16 utils folder, reverted training file
